Add files via upload
Example scene that looks for one of two devices via the Office BLE Presence sensor and checks the workday sensor (separate Home Assistant add-on).  If either Bluetooth device is seen and the workday sensor is "on" (meaning it's a US workday), then it triggers my "scene.office_video_calls" Home Assistant scene which turns on several lights and sets them to the proper color temperature and brightness.